test: mvc 테스트 추가

+++ src/main/java/com/sprint/monew/domain/interest/subscription/SubscriptionDto.java
@@ -9,7 +9,7 @@ public record SubscriptionDto(
     UUID interestId,
     String interestName,
     List<String> interestKeywords,
-    int interestSubscriberCount,
+    long interestSubscriberCount,
     Instant createdAt
 ) {
